simplify kotlin native module update
